+++ LibCarla/source/carla/sensor/data/ActorDynamicState.h
@@ -90,16 +90,27 @@ namespace detail {
 #pragma pack(pop)
 
 #pragma pack(push, 1)
+
   struct TrafficLightData {
     TrafficLightData() = default;
 
-    rpc::TrafficLightState state;
+    char sign_id[32u];
     float green_time;
     float yellow_time;
     float red_time;
     float elapsed_time;
-    bool time_is_frozen;
     uint32_t pole_index;
+    bool time_is_frozen;
+    rpc::TrafficLightState state;
+  };
+#pragma pack(pop)
+
+#pragma pack(push, 1)
+
+  struct TrafficSignData {
+    TrafficSignData() = default;
+
+    char sign_id[32u];
   };
 #pragma pack(pop)
 } // namespace detail
@@ -121,6 +132,7 @@ namespace detail {
 
     union TypeDependentState {
       detail::TrafficLightData traffic_light_data;
+      detail::TrafficSignData traffic_sign_data;
       detail::VehicleData vehicle_data;
       detail::PackedWalkerControl walker_control;
     } state;
@@ -128,27 +140,12 @@ namespace detail {
 
 #pragma pack(pop)
 
-static_assert(
-    sizeof(ActorDynamicState) == 93u,
+ static_assert(
+    sizeof(ActorDynamicState) == 118u,
     "Invalid ActorDynamicState size! "
     "If you modified this class please update the size here, else you may "
     "comment this assert, but your platform may have compatibility issues "
     "connecting to other platforms.");
-
-#pragma pack(push, 1)
-
-/// Dynamic state of a component of an actor at a certain frame.
-struct ComponentDynamicState {
-
-  // World transform
-  geom::Transform transform;
-
-  union TypeDependentState {
-    detail::TrafficLightData traffic_light_data;
-  } state;
-
-};
-#pragma pack(pop)
 
 } // namespace data
 } // namespace sensor